+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/rtc/RTCComponentFactory.kt
@@ -127,8 +127,12 @@ class RTCComponentFactory(private val option: SoraMediaOption,
         }
 
         return JavaAudioDeviceModule.builder(appContext)
-                .setUseHardwareAcousticEchoCanceler(option.audioOption.useHardwareAcousticEchoCanceler)
-                .setUseHardwareNoiseSuppressor(option.audioOption.useHardwareNoiseSuppressor)
+                .setUseHardwareAcousticEchoCanceler(
+                        JavaAudioDeviceModule.isBuiltInAcousticEchoCancelerSupported()
+                                && option.audioOption.useHardwareAcousticEchoCanceler)
+                .setUseHardwareNoiseSuppressor(
+                        JavaAudioDeviceModule.isBuiltInNoiseSuppressorSupported()
+                                && option.audioOption.useHardwareNoiseSuppressor)
                 .setAudioRecordErrorCallback(audioRecordErrorCallback)
                 .setAudioTrackErrorCallback(audioTrackErrorCallback)
                 .createAudioDeviceModule()